% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion_metrics.R
\name{confusion_metrics}
\alias{confusion_metrics}
\title{Confusion matrix metrics for binary classification}
\usage{
confusion_metrics(y_true, p_hat, cutoff = 0.5)
}
\arguments{
\item{y_true}{True 0/1 outcomes}

\item{p_hat}{Predicted probabilities}

\item{cutoff}{Probability cutoff (default 0.5)}
}
\value{
Data frame with accuracy, sensitivity, specificity, FDR, DOR
}
\description{
Confusion matrix metrics for binary classification
}
