% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{multi_path_AIC_pipeline}
\alias{multi_path_AIC_pipeline}
\title{Algorithm 4: Complete Multi-Path AIC Pipeline}
\usage{
multi_path_AIC_pipeline(
  X,
  y,
  model_type = c("gaussian", "binomial"),
  K = NULL,
  eps = 1e-06,
  delta = 2,
  L = 50,
  B = 50,
  resample_type = c("bootstrap", "subsample"),
  m = NULL,
  Delta = 2,
  tau = 0.6
)
}
\arguments{
\item{X}{data frame of predictors}

\item{y}{response vector}

\item{model_type}{"gaussian" or "binomial"}

\item{K, }{eps, delta, L Algorithm 1 parameters}

\item{B, }{resample_type, m Algorithm 2 parameters}

\item{Delta, }{tau Algorithm 3 parameters}
}
\value{
list with:
  - mp_full: Algorithm 1 output
  - stability_pi: Algorithm 2 output
  - plausible: Algorithm 3 output
}
\description{
Algorithm 4: Complete Multi-Path AIC Pipeline
}
\keyword{internal}
