% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R, R/plausible_models.R
\name{select_plausible_models}
\alias{select_plausible_models}
\title{Algorithm 3: Select plausible, stable models}
\usage{
select_plausible_models(mp_full, stability_pi, Delta = 2, tau = 0.6)

select_plausible_models(mp_full, stability_pi, Delta = 2, tau = 0.6)
}
\arguments{
\item{mp_full}{Output from \code{build_paths()}}

\item{stability_pi}{Named numeric stability vector}

\item{Delta}{AIC window}

\item{tau}{Stability threshold}
}
\value{
Data frame of plausible models:
  columns = key, AIC, AIC_diff, size, vars, avg_stability

Data frame of plausible models
}
\description{
Algorithm 3: Select plausible, stable models

Internal: select plausible, stable models
}
\keyword{internal}
