% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_models.R
\name{plausible_models}
\alias{plausible_models}
\title{Plausible model selection using AIC and stability}
\usage{
plausible_models(forest, pi, Delta = 2, tau = 0.6)
}
\arguments{
\item{forest}{Object returned by \code{build_paths()}}

\item{pi}{Named numeric vector of stability scores}

\item{Delta}{AIC tolerance window}

\item{tau}{Minimum average stability}
}
\value{
Data frame of plausible models
}
\description{
Plausible model selection using AIC and stability
}
