% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plausible_models.R
\name{select_plausible_models}
\alias{select_plausible_models}
\title{Internal: select plausible, stable models}
\usage{
select_plausible_models(mp_full, stability_pi, Delta = 2, tau = 0.6)
}
\arguments{
\item{mp_full}{Output from \code{build_paths()}}

\item{stability_pi}{Named numeric stability vector}

\item{Delta}{AIC window}

\item{tau}{Stability threshold}
}
\value{
Data frame of plausible models
}
\description{
Internal: select plausible, stable models
}
\keyword{internal}
