% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{stability}
\alias{stability}
\title{Resampling-based variable stability}
\usage{
stability(
  X,
  y,
  family = c("gaussian", "binomial"),
  B = 50,
  resample = c("bootstrap", "subsample"),
  m = NULL,
  K = NULL,
  eps = 1e-06,
  delta = 2,
  L = 50
)
}
\arguments{
\item{X}{Data frame of predictors}

\item{y}{Response vector}

\item{family}{"gaussian" or "binomial"}

\item{B}{Number of resamples}

\item{resample}{"bootstrap" or "subsample"}

\item{m}{Subsample size if using subsample}

\item{K}{Max model size (steps)}

\item{eps}{Minimum AIC improvement}

\item{delta}{AIC tolerance for near-ties}

\item{L}{Max models to retain per step}
}
\value{
Named numeric vector of stability scores between 0 and 1.
}
\description{
Resampling-based variable stability
}
